+++ packages/balances-evm-erc20/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @talismn/balances-evm-erc20
 
+## 0.3.2
+
+### Patch Changes
+
+- @talismn/balances@0.3.2
+
+## 0.3.1
+
+### Patch Changes
+
+- 8adc7f06: feat: switched build tool to preconstruct
+- Updated dependencies [8adc7f06]
+  - @talismn/chaindata-provider@0.4.1
+  - @talismn/balances@0.3.1
+  - @talismn/util@0.1.6
+
 ## 0.3.0
 
 ### Minor Changes